Fix export to cloud storage && fix enqueuing dependents after canceling rq job

+++ cvat/apps/engine/views.py
@@ -7,7 +7,7 @@ import os
 import os.path as osp
 from PIL import Image
 from types import SimpleNamespace
-from typing import Optional, Any, Dict, List, cast
+from typing import Optional, Any, Dict, List, cast, Callable
 import traceback
 import textwrap
 from copy import copy
@@ -20,9 +20,9 @@ import django_rq
 from django.conf import settings
 from django.contrib.auth.models import User
 from django.db import IntegrityError, transaction
-from django.db.models import Count
+from django.db.models import Count, Model
 from django.db.models.query import Prefetch
-from django.http import HttpResponse, HttpResponseNotFound, HttpResponseBadRequest
+from django.http import HttpResponse, HttpRequest, HttpResponseNotFound, HttpResponseBadRequest
 from django.utils import timezone
 
 from drf_spectacular.types import OpenApiTypes
@@ -42,7 +42,7 @@ from rest_framework.settings import api_settings
 
 import cvat.apps.dataset_manager as dm
 import cvat.apps.dataset_manager.views  # pylint: disable=unused-import
-from cvat.apps.engine.cloud_provider import db_storage_to_storage_instance
+from cvat.apps.engine.cloud_provider import db_storage_to_storage_instance, download_file_from_bucket, export_resource_to_cloud_storage
 from cvat.apps.dataset_manager.bindings import CvatImportError
 from cvat.apps.dataset_manager.serializers import DatasetFormatsSerializer
 from cvat.apps.engine.frame_provider import FrameProvider
@@ -71,7 +71,8 @@ from utils.dataset_manifest import ImageManifestManager
 from cvat.apps.engine.utils import (
     av_scan_paths, process_failed_job, configure_dependent_job_to_download_from_cs,
     parse_exception_message, get_rq_job_meta, get_import_rq_id,
-    import_resource_with_clean_up_after, sendfile, define_dependent_job, get_rq_lock_by_user
+    import_resource_with_clean_up_after, sendfile, define_dependent_job, get_rq_lock_by_user,
+    build_annotations_file_name,
 )
 from cvat.apps.engine import backup
 from cvat.apps.engine.mixins import PartialUpdateModelMixin, UploadMixin, AnnotationMixin, SerializeMixin
@@ -292,7 +293,7 @@ class ProjectViewSet(viewsets.GenericViewSet, mixins.ListModelMixin,
                 location=OpenApiParameter.QUERY, type=OpenApiTypes.STR, required=False),
             OpenApiParameter('filename', description='Desired output file name',
                 location=OpenApiParameter.QUERY, type=OpenApiTypes.STR, required=False),
-            OpenApiParameter('action', description='Used to start downloading process after annotation file had been created',
+            OpenApiParameter('action', description='Used to start downloading process locally after annotation file has been created',
                 location=OpenApiParameter.QUERY, type=OpenApiTypes.STR, required=False, enum=['download', 'import_status']),
             OpenApiParameter('location', description='Where need to save downloaded dataset',
                 location=OpenApiParameter.QUERY, type=OpenApiTypes.STR, required=False,
@@ -456,7 +457,7 @@ class ProjectViewSet(viewsets.GenericViewSet, mixins.ListModelMixin,
                 location=OpenApiParameter.QUERY, type=OpenApiTypes.STR, required=True),
             OpenApiParameter('filename', description='Desired output file name',
                 location=OpenApiParameter.QUERY, type=OpenApiTypes.STR, required=False),
-            OpenApiParameter('action', description='Used to start downloading process after annotation file had been created',
+            OpenApiParameter('action', description='Used to start downloading process locally after annotation file has been created',
                 location=OpenApiParameter.QUERY, type=OpenApiTypes.STR, required=False, enum=['download']),
             OpenApiParameter('location', description='Where need to save downloaded dataset',
                 location=OpenApiParameter.QUERY, type=OpenApiTypes.STR, required=False,
@@ -1233,7 +1234,7 @@ class TaskViewSet(viewsets.GenericViewSet, mixins.ListModelMixin,
             OpenApiParameter('filename', description='Desired output file name',
                 location=OpenApiParameter.QUERY, type=OpenApiTypes.STR, required=False),
             OpenApiParameter('action', location=OpenApiParameter.QUERY,
-                description='Used to start downloading process after annotation file had been created',
+                description='Used to start downloading process locally after annotation file has been created',
                 type=OpenApiTypes.STR, required=False, enum=['download']),
             OpenApiParameter('location', description='Where need to save downloaded dataset',
                 location=OpenApiParameter.QUERY, type=OpenApiTypes.STR, required=False,
@@ -1484,7 +1485,7 @@ class TaskViewSet(viewsets.GenericViewSet, mixins.ListModelMixin,
             OpenApiParameter('filename', description='Desired output file name',
                 location=OpenApiParameter.QUERY, type=OpenApiTypes.STR, required=False),
             OpenApiParameter('action', location=OpenApiParameter.QUERY,
-                description='Used to start downloading process after annotation file had been created',
+                description='Used to start downloading process locally after annotation file has been created',
                 type=OpenApiTypes.STR, required=False, enum=['download']),
             OpenApiParameter('use_default_location', description='Use the location that was configured in task to export annotations',
                 location=OpenApiParameter.QUERY, type=OpenApiTypes.BOOL, required=False,
@@ -1668,7 +1669,7 @@ class JobViewSet(viewsets.GenericViewSet, mixins.ListModelMixin, mixins.CreateMo
             OpenApiParameter('filename', description='Desired output file name',
                 location=OpenApiParameter.QUERY, type=OpenApiTypes.STR, required=False),
             OpenApiParameter('action', location=OpenApiParameter.QUERY,
-                description='Used to start downloading process after annotation file had been created',
+                description='Used to start downloading process locally after annotation file has been created',
                 type=OpenApiTypes.STR, required=False, enum=['download']),
             OpenApiParameter('location', description='Where need to save downloaded annotation',
                 location=OpenApiParameter.QUERY, type=OpenApiTypes.STR, required=False,
@@ -1845,7 +1846,7 @@ class JobViewSet(viewsets.GenericViewSet, mixins.ListModelMixin, mixins.CreateMo
             OpenApiParameter('filename', description='Desired output file name',
                 location=OpenApiParameter.QUERY, type=OpenApiTypes.STR, required=False),
             OpenApiParameter('action', location=OpenApiParameter.QUERY,
-                description='Used to start downloading process after annotation file had been created',
+                description='Used to start downloading process locally after annotation file has been created',
                 type=OpenApiTypes.STR, required=False, enum=['download']),
             OpenApiParameter('use_default_location', description='Use the location that was configured in the task to export dataset',
                 location=OpenApiParameter.QUERY, type=OpenApiTypes.BOOL, required=False,
@@ -2801,12 +2802,6 @@ def rq_exception_handler(rq_job, exc_type, exc_value, tb):
 
     return True
 
-def _download_file_from_bucket(db_storage, filename, key):
-    storage = db_storage_to_storage_instance(db_storage)
-
-    with storage.download_fileobj(key) as data, open(filename, 'wb+') as f:
-        f.write(data.getbuffer())
-
 def _import_annotations(request, rq_id_template, rq_func, db_obj, format_name,
                         filename=None, location_conf=None, conv_mask_to_poly=True):
 
@@ -2879,7 +2874,7 @@ def _import_annotations(request, rq_id_template, rq_func, db_obj, format_name,
                 dependent_job = configure_dependent_job_to_download_from_cs(
                     queue=queue,
                     rq_id=rq_id,
-                    rq_func=_download_file_from_bucket,
+                    rq_func=download_file_from_bucket,
                     db_storage=db_storage,
                     filename=filename,
                     key=key,
@@ -2928,8 +2923,16 @@ def _import_annotations(request, rq_id_template, rq_func, db_obj, format_name,
 
     return Response(status=status.HTTP_202_ACCEPTED)
 
-def _export_annotations(db_instance, rq_id, request, format_name, action, callback,
-                        filename, location_conf):
+def _export_annotations(
+    db_instance: Model,
+    rq_id: str,
+    request: HttpRequest,
+    format_name: str,
+    action: str,
+    callback: Callable[[int, Optional[str], Optional[str]], str],
+    filename: Optional[str],
+    location_conf: Dict[str, Any]
+):
     if action not in {"", "download"}:
         raise serializers.ValidationError(
             "Unexpected action specified for the request")
@@ -2945,60 +2948,59 @@ def _export_annotations(db_instance, rq_id, request, format_name, action, callba
     queue = django_rq.get_queue(settings.CVAT_QUEUES.EXPORT_DATA.value)
     rq_job = queue.fetch_job(rq_id)
 
+    location = location_conf.get('location')
+    if location not in Location.list():
+        raise serializers.ValidationError(
+            f"Unexpected location {location} specified for the request"
+        )
+
+    last_instance_update_time = timezone.localtime(db_instance.updated_date)
+    if isinstance(db_instance, Project):
+        tasks_update = list(map(lambda db_task: timezone.localtime(db_task.updated_date), db_instance.tasks.all()))
+        last_instance_update_time = max(tasks_update + [last_instance_update_time])
+
+    timestamp = datetime.strftime(last_instance_update_time, "%Y_%m_%d_%H_%M_%S")
+    are_annotations_exported = 'annotations' in rq_id
+
     if rq_job:
-        last_instance_update_time = timezone.localtime(db_instance.updated_date)
-        if isinstance(db_instance, Project):
-            tasks_update = list(map(lambda db_task: timezone.localtime(db_task.updated_date), db_instance.tasks.all()))
-            last_instance_update_time = max(tasks_update + [last_instance_update_time])
         rq_request = rq_job.meta.get('request', None)
         request_time = rq_request.get('timestamp', None) if rq_request else None
         if request_time is None or request_time < last_instance_update_time:
-            rq_job.cancel()
+            # in case the server is configured with ONE_RUNNING_JOB_IN_QUEUE_PER_USER
+            # we have to enqueue dependent jobs after canceling one
+            rq_job.cancel(enqueue_dependents=settings.ONE_RUNNING_JOB_IN_QUEUE_PER_USER)
             rq_job.delete()
         else:
             if rq_job.is_finished:
-                file_path = rq_job.return_value()
-                if action == "download" and osp.exists(file_path):
+                if location == Location.CLOUD_STORAGE:
                     rq_job.delete()
+                    return Response(status=status.HTTP_200_OK)
 
-                    timestamp = datetime.strftime(last_instance_update_time,
-                        "%Y_%m_%d_%H_%M_%S")
-                    filename = filename or \
-                        "{}_{}-{}-{}{}".format(
-                            db_instance.__class__.__name__.lower(),
-                            db_instance.name if isinstance(db_instance, (Task, Project)) else db_instance.id,
-                            timestamp, format_name, osp.splitext(file_path)[1]
-                        ).lower()
+                elif location == Location.LOCAL:
+                    file_path = rq_job.return_value()
 
-                    # save annotation to specified location
-                    location = location_conf.get('location')
-                    if location == Location.LOCAL:
-                        return sendfile(request, file_path, attachment=True,
-                            attachment_filename=filename)
-                    elif location == Location.CLOUD_STORAGE:
-                        try:
-                            storage_id = location_conf['storage_id']
-                        except KeyError:
-                            return HttpResponseBadRequest(
-                                'Cloud storage location was selected as the destination,'
-                                ' but cloud storage id was not specified')
-                        db_storage = get_cloud_storage_for_import_or_export(
-                            storage_id=storage_id, request=request,
-                            is_default=location_conf['is_default'])
-                        storage = db_storage_to_storage_instance(db_storage)
+                    if not file_path:
+                        return Response('A result for exporting job was not found for finished RQ job', status=status.HTTP_500_INTERNAL_SERVER_ERROR)
+                    elif not osp.exists(file_path):
+                        return Response('Result file does not exists in export cache', status=status.HTTP_500_INTERNAL_SERVER_ERROR)
 
-                        try:
-                            storage.upload_file(file_path, filename)
-                        except (ValidationError, PermissionDenied, NotFound) as ex:
-                            msg = str(ex) if not isinstance(ex, ValidationError) else \
-                                '\n'.join([str(d) for d in ex.detail])
-                            return Response(data=msg, status=ex.status_code)
-                        return Response(status=status.HTTP_200_OK)
-                    else:
-                        raise NotImplementedError()
+                    if action == "download":
+                        filename = filename or \
+                            build_annotations_file_name(
+                                class_name=db_instance.__class__.__name__,
+                                identifier=db_instance.name if isinstance(db_instance, (Task, Project)) else db_instance.id,
+                                timestamp=timestamp,
+                                format_name=format_name,
+                                is_annotation_file=are_annotations_exported,
+                                extension=osp.splitext(file_path)[1]
+                            )
+
+                        rq_job.delete()
+                        return sendfile(request, file_path, attachment=True, attachment_filename=filename)
+
+                    return Response(status=status.HTTP_201_CREATED)
                 else:
-                    if osp.exists(file_path):
-                        return Response(status=status.HTTP_201_CREATED)
+                    raise NotImplementedError(f"Export to {location} location is not implemented yet")
             elif rq_job.is_failed:
                 exc_info = rq_job.meta.get('formatted_exception', str(rq_job.exc_info))
                 rq_job.delete()
@@ -3022,10 +3024,33 @@ def _export_annotations(db_instance, rq_id, request, format_name, action, callba
     ttl = TTL_CONSTS[db_instance.__class__.__name__.lower()].total_seconds()
     user_id = request.user.id
 
+    func = callback if location == Location.LOCAL else export_resource_to_cloud_storage
+    func_args = (db_instance.id, format_name, server_address)
+
+    if location == Location.CLOUD_STORAGE:
+        try:
+            storage_id = location_conf['storage_id']
+        except KeyError:
+            raise serializers.ValidationError(
+                'Cloud storage location was selected as the destination,'
+                ' but cloud storage id was not specified')
+
+        db_storage = get_cloud_storage_for_import_or_export(
+            storage_id=storage_id, request=request,
+            is_default=location_conf['is_default'])
+        filename_pattern = build_annotations_file_name(
+            class_name=db_instance.__class__.__name__,
+            identifier=db_instance.name if isinstance(db_instance, (Task, Project)) else db_instance.id,
+            timestamp=timestamp,
+            format_name=format_name,
+            is_annotation_file=are_annotations_exported,
+        )
+        func_args = (db_storage, filename, filename_pattern, callback) + func_args
+
     with get_rq_lock_by_user(queue, user_id):
         queue.enqueue_call(
-            func=callback,
-            args=(db_instance.id, format_name, server_address),
+            func=func,
+            args=func_args,
             job_id=rq_id,
             meta=get_rq_job_meta(request=request, db_obj=db_instance),
             depends_on=define_dependent_job(queue, user_id, rq_id=rq_id),
@@ -3090,7 +3115,7 @@ def _import_project_dataset(request, rq_id_template, rq_func, db_obj, format_nam
             dependent_job = configure_dependent_job_to_download_from_cs(
                 queue=queue,
                 rq_id=rq_id,
-                rq_func=_download_file_from_bucket,
+                rq_func=download_file_from_bucket,
                 db_storage=db_storage,
                 filename=filename,
                 key=key,
